aes: fourteen 24-bit addresses that were being cut to sixteen
The policy change is four lines.  Surviving it was the work.

ob_spec, te_ptext, te_ptmplt, te_pvalid, ib_ptext, ib_pdata, mn_tree and
fd_addr are LONGs holding ADDRESSES.  While every resource lived in bank
$00, taking the low sixteen bits of one was correct, and a great deal of
code did, in the plainest way.  Fourteen places, in two flavours.

AN EXPLICIT (uint16_t), stating an assumption that had just stopped
being true: nine in src/desk/, four in clock.c.  These were findable --
the clang warnings from tools/ccdep.sh name them.

AND A SILENT ONE, which nothing warns about.  wind_set(WF_NAME) and
WF_INFO in deskwin.c passed a title's low word and a hardcoded 0 for the
high word; WF_NEWDESK did the same in the gem.h macro and in wind.c.
The call takes two words and was given two words, one of them wrong.

Two were their own shape.  menu_register had been bouncing its string
through abi.c's 64-byte str_scratch, which is right for a call that
reads its argument and returns -- but the AES keeps a registered name
for the life of the machine and str_used winds back to zero on the next
call.  It stores all 24 bits now.  And w_ptext used to COPY a far title
into a 41-byte near buffer: that capped a far title at forty characters
while a near one had none, and the desktop's own w_name is fifty bytes,
so a deep path drew short from the day G moved to far memory -- and
tools/aesref.py never modelled the cap, so model and target had quietly
stopped agreeing.  It assigns the address now, gl_nbuf/gl_ibuf are
gone (+82 bytes of LoRAM), and objc_draw was already reading a G_TEXT's
string through a far pointer, which is the same path a frame takes.

app.c exports app_far so a gate can find a large-data program's globals,
and abi.c gains gem_badop -- which opcode the last refused call was in.
It earned itself at once: shel_write was refusing "" and "\0" from
SHW_SHUTDOWN, far literals in a large-data program, and gem_badop named
opcode 121 instead of leaving it to be hunted.

+++ src/apps/clock.c
@@ -39,9 +39,9 @@ static WORD day, month, year;               /* year is the century's two */
 
 static char *field(WORD obj)
 {
-    TEDINFO *ted = (TEDINFO *)(uint16_t)tree[obj].ob_spec.index;
+    TEDINFO *ted = (TEDINFO *)(uint32_t)tree[obj].ob_spec.index;
 
-    return (char *)(uint16_t)ted->te_ptext;
+    return (char *)(uint32_t)ted->te_ptext;
 }
 
 /* Three numbers into the runs of underscores of a field's template, in
@@ -49,8 +49,8 @@ static char *field(WORD obj)
  * order, this decides only what the numbers are. */
 static void fill(WORD obj, WORD a, WORD b, WORD c)
 {
-    TEDINFO *ted = (TEDINFO *)(uint16_t)tree[obj].ob_spec.index;
-    const char *t = (const char *)(uint16_t)ted->te_ptmplt;
+    TEDINFO *ted = (TEDINFO *)(uint32_t)tree[obj].ob_spec.index;
+    const char *t = (const char *)(uint32_t)ted->te_ptmplt;
     char *d = field(obj);
     WORD v[3], n = 0, i = 0;
